chore(hash-stream-node): move fileStreamHasher to own file
